flask-tdd-docker/requirements.txt
Flask==2.3.3
flask-restx==1.2.0
pytest==7.4.3
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
pymysql==1.1.0
